Extract `Deleg` ExecSpecRule and `SpecTranslate` instances
